India's Demonetization Analysis
The objective of this Case study is to analyze sentiments from twitter of a highly spoken-of event that took place on 8th November 2016 in India. To combat the black money problem in the country, the government decided to demonetize two bills of the highest denominations. We will analyze on how the users reacted to the event using a dictionary and further on, we will visualize the results obtained.