Machine Learning with Pyspark
PySpark is a great language for performing exploratory data analysis and building machine learning pipelines. MLlib is a Spark implementation of useful machine learning (ML) functionality. The objective of this case study is to become familiar with Pyspark. We will load the data in Pyspark and build a machine learning model using the K-Means algorithm.